Fixes completed

+++ src/main/java/lejos/hardware/gps/GGASentence.java
@@ -118,7 +118,7 @@ public class GGASentence extends NMEASentence {
 //			}
 //		}
 		
-		return (double)(degrees + minutes * (double)(1.0 / 60.0));
+		return degrees + minutes * (1.0 / 60.0);
 	}
 	
 	/*
